feat(internal): add some feature flags for new auth libs
We will use the internaloption to start to enable new auth lib
for a small number of clients in the future. The envvar will be
undocumented and used exclusively in our own testing environments.
We will enable it on all of our repos as the first part of
functional testing that all existing integration tests continue to
work as normal when routing through the new auth layers.