feat: expose experimental universe-related options (#2264)
This PR adds the first step in plumbing a universe option into clients
by first exposing it as part of the option package.

It exposes a `WithUniverse(string)` as a public option, and `WithDefaultUniverse(string)`
as an internal option that can be used by generated clients.  It also augments existing
tests.

It does not yet include any behavioral changes.