feat: add hooks in for new auth library (#2228)
All code that calls into the old base auth layer can not optionally call into the new stack. Where possible, packages call into new auth packages directly. As a fallback, and for cases where we don't have 1-1 replacements the code in internal should ensure that at least the ADC logic from the new auth stack is called into.

Flipping the environment variable on for this repo so we make sure that all tests currently written pass without any code changes. This should give confidence that when options are in place to call into the new stack code continues to work as is without any code changes.