fix(gensupport): fix transferChunk race condition by returning response with non-cancelled context. (#3258)
The previous implementation used context.WithTimeout to manage timeouts for each chunk upload and cancelled the upload when the response was received. This introduced a race condition when the response was not yet processed by caller and context was cancelled.

This fix removes the race condition by doing following things:

- Starts a timer and the upload request in a separate goroutine.
- Uses a select statement to wait for either the upload to complete, the timer to fire, or the context to be cancelled.
- Explicitly cancels the upload request if the timer fires first, preventing the race condition by not cancelling the context when response is succeess (which waits for caller to process the body) and ensuring resources are cleaned up correctly.
- Renamed two variable to make code more readable.

Internal Bug: 435359905